# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Core API",
        "version": "1.0.0"
    },
    "basePath": "/api",
    "paths": {
        "/list": {
            "get": {
                "summary": "Gets list of messages",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Read status",
                        "required": false,
                        "type": "string",
                        "enum": [
                            "read",
                            "unread"
                        ]
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "From",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "How many records should be returned",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/listArchived": {
            "get": {
                "summary": "Gets list of archived messages",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Read status",
                        "required": false,
                        "type": "string",
                        "enum": [
                            "read",
                            "unread"
                        ]
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "From",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "How many records should be returned",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/read/{uid}": {
            "put": {
                "summary": "Mark a message as read",
                "parameters": [
                    {
                        "name": "uid",
                        "in": "path",
                        "description": "Id of the message",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/archive/{uid}": {
            "put": {
                "summary": "Mark a message as read",
                "parameters": [
                    {
                        "name": "uid",
                        "in": "path",
                        "description": "Id of the message",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/message/{uid}": {
            "get": {
                "summary": "Display a message",
                "parameters": [
                    {
                        "name": "uid",
                        "in": "path",
                        "description": "Id of the message",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        }
    },
    "definitions": {}
}